Fix data overlow bug in maxMsgSize computation (#3782)
When we compute the 90% of this we were doing
an integer value of 64 MB times 90 which overflows
and results in a value of 17MB computed max size.
This limits the amount of data that is batched and has performance
ramifications.
fix long in BaseLogEntryReader & BaseSnapshotReader
NullPointerException fix in LogReplicationClientServerRouter

+++ infrastructure/src/main/java/org/corfudb/infrastructure/ServerContext.java
@@ -288,7 +288,7 @@ public class ServerContext implements AutoCloseable {
      *
      * @return max data message size
      */
-    public int getLogReplicationMaxDataMessageSize() {
+    public long getLogReplicationMaxDataMessageSize() {
         String val = getServerConfig(String.class, "--max-replication-data-message-size");
         return val == null ? DEFAULT_MAX_DATA_MSG_SIZE : Integer.parseInt(val);
     }
